Dennis:
Bug description:

Create an empty project.
Add an assignment step "Assign"
Add an assignment set "Assign1" after "Assign".
Add a scope anywhere.
Now, drag an drop the "Assign" step into the scope.   This works fine.
Now drag and drop the "Assign1" step into the scope AFTER "assign".   The
"Assign1" step is lost, and the scope now contains a Sequence, and only
"Assign".

If you drop "Assign1" BEFORE "Assign", it is added correctly.  Only trying to
place it after causes it to be lost.

It seems that a Sequence is also added, and it seems adding both the sequence
and assignment is a problem.  If the sequence is already there, then the
problem does not occur.

Fixed:
- made Sequence correctly adopt/orphan children instead of get patched by patchDom (+regression test)
- fixed isSingleActivityContainer to return true for Scope (+regression test)
- added method to get all contained activities from DOM element for Sequence adopt (+test)