Dennis:
Bug description:

Create a new project
Create Assign

Under details, set the "from" to a Fixed value, and enter the following:

<rawFilter><![CDATA[(objectClass=*)]]>
<other/><data/></rawFi

The rawFilter tag is intentionally truncated.   Now, Click on the receive step.
A pop up will say that the data is not valid XML.  Click OK.
Now if you save the BPEL, close, and reopen the file, the partnerlinks have
been removed.

If you drop "Assign1" BEFORE "Assign", it is added correctly.  Only trying to
place it after causes it to be lost.

It seems that a Sequence is also added, and it seems adding both the sequence
and assignment is a problem.  If the sequence is already there, then the
problem does not occur.

Fixed:
all process children should be set to null when document is changed to re-read
all DOM elements